ignore contested greengenes taxons that throw off the rare taxon list

+++ scripts/generate_otu_signifigance_tables_AGP.py
@@ -130,6 +130,11 @@ def calculate_tax_rank_1(sample, population, taxa, rare_threshold=0.1, \
 
         elif sample_bin[idx] == 1 and \
              population_count[idx] < num_samples*rare_threshold:
+
+            # ignore contested groupings
+            if '[' in taxon:
+                continue
+
             rare.append(taxon)
             remove_index.append(idx)
